requirements.txt
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
mysqlclient==2.1.1
channels==4.0.0
daphne==4.0.0
pytest==7.3.1
pytest-django==4.5.2